Chester County Democratic Committee - Voter Resource App
This is a hybrid Mobile Application built using the Ionic Framework
(v2), AngularJS and Google Civic Information API. This project is
intended for reference only and will not be maintained. If I create a
new version that will be maintained I will link to it. I hope this code
base will help other developers exploring Ionic or the Civic Info API.